% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{cmp_leave_one_time_bin_out_cross_validation}
\alias{cmp_leave_one_time_bin_out_cross_validation}
\title{Compute network flow solutions for different values of mu and where one time bin is left out
The output is a list where each element is the return value of the function solve_piecewise_linear_mosek_network for the different input parameter values of mu and time_bin_out}
\usage{
cmp_leave_one_time_bin_out_cross_validation(
  df_cv_parameter,
  mct,
  mct_t_sd = NULL
)
}
\arguments{
\item{df_cv_parameter}{data.frame with column mu and column time_bin_out}

\item{mct}{mct object that is the output}

\item{mc_t_sd}{optional}
}
\value{
list where each element is the return value of the function solve_piecewise_linear_mosek_network for the different input parameter values of mu and time_bin_out
}
\description{
Compute network flow solutions for different values of mu and where one time bin is left out
The output is a list where each element is the return value of the function solve_piecewise_linear_mosek_network for the different input parameter values of mu and time_bin_out
}
