% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{initialize,tgMCTNetwork-method}
\alias{initialize,tgMCTNetwork-method}
\title{Construct a metacell time network}
\usage{
\S4method{initialize}{tgMCTNetwork}(.Object, mc_t, temporal_bin_time = NULL, metacell_names = NULL)
}
\arguments{
\item{mc_t}{distribution of metacells (rows) over time points (cols)}
}
\description{
Construct a metacell time network
}
