% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mc2_new_mctnetwork}
\alias{mc2_new_mctnetwork}
\title{Generate a new network in scdb}
\usage{
mc2_new_mctnetwork(mc_t, temporal_bin_time = NULL, metacell_names = NULL)
}
\arguments{
\item{mc_t}{matrix distribution of metacells (rows) over time points (cols)}

\item{temporal_bin_time}{data.frame with columns bin and t giving for each temporal bin its actual time}

\item{metacell_names}{vector with metacell names corresponding to rownames of mc_t}
}
\description{
This constructs a metacell time network object with mc_t entry and metacell_names and temporal_bin_time_entry
}
