% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mcell_mctnet_from_mgraph}
\alias{mcell_mctnet_from_mgraph}
\title{Construct and initialize a mct time network from a graph and leak table}
\usage{
mcell_mctnet_from_mgraph(
  net_id,
  mgraph_id,
  cell_time,
  leak_fn,
  t_exp = 1,
  T_cost = 100000,
  capacity_var_factor = NULL,
  off_capacity_cost1 = 1,
  off_capacity_cost2 = 1000,
  k_norm_ext_cost = 1,
  k_ext_norm_cost = 1,
  k_ext_ext_cost = 1
)
}
\arguments{
\item{net_id}{id of scdb network object ot be added}

\item{mgraph_id}{metacell manfold graph object}

\item{cell_time}{assigning of time (discrete) to cells}

\item{leak_fn}{a table holding the leak parameter per metacell. This should be substituted with a metacell metadata field. Leaks are the fracitons of frequency loss per mc per time step. All zero if you don't know what to expect.}

\item{t_exp}{the time period for the matrix expoenential used when computing manifold costs. This is default to 1, and should usually stay this way}

\item{T_cost}{the threshold of manifold edges costs that are removed from the network to make the resulting optimization problem smaller.}

\item{capacity_var_factor}{optionally a vector specifying how tight should the capacity constraint on each metacell is modeled.}

\item{off_capacity_cost1}{the slope of the piecewise linear capacity cost around 1. Typically a smaller number at an order of 1}

\item{off_capacity_cost2}{the slope of the piecewise linear cpaacity cost away from 1 - tpyically a large number representing strong constraint for not meeting the capacity constraints at all (either too little or too much flow)}

\item{k_norm_ext_cost}{cost of flow from normal to extneded capacity (obselete should be removed, keep it as 1)}

\item{k_ext_norm_cost}{cost of flow from extneded to normal capacity (obselete should be removed, keep it as 1)}

\item{k_ext_ext_cost}{cost of flow from extneded to extended capacity (obselete should be removed, keep it as 1)}
}
\description{
This constructs a meta cell time network object from an mc and mgraph objects and assignments of cells to graphs
}
