% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mcell_new_mctnetflow}
\alias{mcell_new_mctnetflow}
\title{Generate a new flow object in scdb}
\usage{
mcell_new_mctnetflow(
  flow_id,
  net_id,
  scdb_dir,
  init_mincost = F,
  flow_tolerance = 0.01,
  max_flow_tolerance = 0.05,
  flows = NULL
)
}
\arguments{
\item{flow_id}{id of scdb network object ot be added}

\item{net_id}{id of scdb mc time network object}

\item{init_mincost}{TRUE if you want to compute flows using the mincost algorith}

\item{flows}{flows per network edge (NULL by default, if not NULL it will supress the init_mincost flag)}
}
\description{
This constructs a flow object and put in the flow scdb. Flow themselves are NULL by default, use gen_mincost to compute them.
}
