% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mctnetflow_comp_propagation}
\alias{mctnetflow_comp_propagation}
\title{Compute matrices for forward and backwrd propagation using the flows}
\usage{
mctnetflow_comp_propagation(mcf)
}
\arguments{
\item{mcf}{mcf flow object}
}
\value{
an updated mcf object
}
\description{
Compute matrices for forward and backwrd propagation using the flows
}
