% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mctnetflow_gen_e_gmt_p}
\alias{mctnetflow_gen_e_gmt_p}
\title{This generate two large matrices showing expression of genes per metacell ad time point, as well as the expression of the gene in inferred ancestral states given the flow model}
\usage{
mctnetflow_gen_e_gmt_p(mcf_id, mat_id, genes)
}
\arguments{
\item{mcf_id}{mcf network object}

\item{mat_id}{umi matrix object}

\item{genes}{list of gene names}
}
\value{
A matrix on metacells and time with mean umi fraction over the gene module
}
\description{
This generate two large matrices showing expression of genes per metacell ad time point, as well as the expression of the gene in inferred ancestral states given the flow model
}
