% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mctnetflow_gen_gmod_mc_t}
\alias{mctnetflow_gen_gmod_mc_t}
\title{Given a list of genes and a matrix, this compute mean umi per metacell (rows) and time (column).}
\usage{
mctnetflow_gen_gmod_mc_t(mcf_id, mat_id, genes, min_percentile = 0.05)
}
\arguments{
\item{mcf_id}{flow object id}

\item{mat_id}{umi matrix object}

\item{genes}{list of gene names}

\item{min_percentile}{percentile value to use as minimum threshold (this is used to avoid very small values and regularize color scale, default 0.05)}
}
\value{
A matrix on metacells and time with mean umi fraction over the gene module
}
\description{
Given a list of genes and a matrix, this compute mean umi per metacell (rows) and time (column).
}
