% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mctnetflow_gen_mincost}
\alias{mctnetflow_gen_mincost}
\title{Compute mincost flow and update the mct objects with flows and 
inferred mc capacity per time}
\usage{
mctnetflow_gen_mincost(mcf, flow_tolerance = 0.01)
}
\arguments{
\item{mcf}{flow object}

\item{flow_tolerance}{how much flow we should miss per time}
}
\value{
an updated flow object
}
\description{
Compute mincost flow and update the mct objects with flows and 
inferred mc capacity per time
}
