% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mctnetflow_get_flow_mat}
\alias{mctnetflow_get_flow_mat}
\title{Compute matrix of flows between MCs in a given time point}
\usage{
mctnetflow_get_flow_mat(mcf, mctnet, time)
}
\arguments{
\item{mcf}{flow object}

\item{mctnet}{mctnet network object}

\item{time}{flows will be computed for the (time,time+1) interval. Time=-1 will generate total flow over all times}
}
\value{
a matrix on metacells
}
\description{
Compute matrix of flows between MCs in a given time point
}
