% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{mctnetflow_order_metacells_by_type_and_time}
\alias{mctnetflow_order_metacells_by_type_and_time}
\title{Orders metacells by their type and within each type by color}
\usage{
mctnetflow_order_metacells_by_type_and_time(
  mc_metadata,
  metadata_field_type,
  metadata_field_time,
  df_type_rank,
  included_metacells = NULL
)
}
\arguments{
\item{mc_metadata}{data.frame with metacell_names as rownames and type and time metadata fields}

\item{metadata_field_type}{column name of type metadata}

\item{metadata_field_time}{column name of time metadata}

\item{df_type_rank}{data.frame with column rank and column metadata_field_type}

\item{included_metacells}{specify this if only a subset of the metacells should be used}
}
\value{
a named vector containing the rank of each metacell
}
\description{
Orders metacells by their type and within each type by color
}
