% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_comp_manifold_costs_from_mgraph}
\alias{mctnetwork_comp_manifold_costs_from_mgraph}
\title{Compute and update manifold costs for a mc time network}
\usage{
mctnetwork_comp_manifold_costs_from_mgraph(
  mct,
  mgraph,
  t_exp = 1,
  T_cost = 1e+05
)
}
\arguments{
\item{mct}{mct network object}

\item{mgraph}{data.frame with columns mc1, mc2 and dist}

\item{t_exp}{time parameter for the markov exponential}

\item{T_cost}{threshold on the mc-mc cost for being included}
}
\value{
an updated mct object with the entry mc_manif_cost updatetd
}
\description{
Compute and update manifold costs for a mc time network
}
