% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_from_mgraph_and_mc_proliferation_rate}
\alias{mctnetwork_from_mgraph_and_mc_proliferation_rate}
\title{Construct and initialize a mct time network from a graph and leak table}
\usage{
mctnetwork_from_mgraph_and_mc_proliferation_rate(
  mgraph,
  mc_t,
  mc_proliferation_rate = NULL,
  temporal_bin_time = NULL,
  t_exp = 1,
  T_cost = 1e+05,
  metacell_names = NULL
)
}
\arguments{
\item{mc_proliferation_rate}{a named vector holding the proliferation rate parameter per metacell, with names being metacell names. All zero if you don't know what to expect. Proliferation rates should be given in real time units e.g. 4 corresponds to approx 4 divisions per time unit. Timbe bin times are specified in the temporal_bin_time parameter.}

\item{temporal_bin_time}{a data.frame with columns bin and t containing the real times of each time bin - used to estimate growth rate per metacell and time point. If not specified, pseudotimes c(1:ncol(mc_t)) are used}

\item{t_exp}{the time period for the matrix expoenential used when computing manifold costs. This is default to 1, and should usually stay this way}

\item{T_cost}{the threshold of manifold edges costs that are removed from the network to make the resulting optimization problem smaller.}

\item{net_id}{id of scdb network object ot be added}

\item{mgraph_id}{metacell manfold graph object}

\item{cell_time}{assigning of time (discrete) to cells}
}
\value{
mct object with computed manifold costs and mctnetwork
}
\description{
This constructs a meta cell time network object from an mc and mgraph objects and assignments of cells to graphs
}
