% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\name{mctnetwork_renormalize_mc_expansion_rates}
\alias{mctnetwork_renormalize_mc_expansion_rates}
\title{normalizes mc_t_expansion_rate given an mc_t distribution such that the total frequency after proliferation per time point still equals one given the frequencies mc_t}
\usage{
mctnetwork_renormalize_mc_expansion_rates(mc_t_expansion_rate, mc_t_freq)
}
\arguments{
\item{mc_t_expansion_rate}{matrix of metacells over time points minus the last time point (i.e. ncol(mc_t_expansion_rate) should equal ncol(mc_t) - 1)}

\item{mc_t}{matrix of metacell frequency over time points}
}
\value{
renormalized mc_t_expansion_rate
}
\description{
normalizes mc_t_expansion_rate given an mc_t distribution such that the total frequency after proliferation per time point still equals one given the frequencies mc_t
}
