% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.flow-package.R
\docType{package}
\name{metacell.flow-package}
\alias{metacell.flow}
\alias{metacell.flow-package}
\title{metacell.flow: Inferring and annotating metacell flows}
\description{
Supporting inference of differentiation flow over metacell manifolds (and innovation of additional scRNA buzzwords)
}
\author{
\strong{Maintainer}: Amos Tanay \email{amos.tanay@weizmann.ac.il}

}
\keyword{internal}
