% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph_logistic.R
\name{metacell2_mgraph_logistic}
\alias{metacell2_mgraph_logistic}
\title{Compute metacell manifold graph using logistic distances and balanced K-nn}
\usage{
metacell2_mgraph_logistic(
  mc2_anndata,
  feature_genes,
  ignore_edges = NULL,
  features_exclude = NULL,
  included_metacells = NULL,
  logist_loc = 1,
  logist_scale = 0.2,
  logist_eps = 4e-05,
  max_d_fold = 3,
  max_deg = 4
)
}
\arguments{
\item{ignore_edges}{provide a data frame with mc1,mc2 pairs of edges to delete manually}

\item{features_exclude}{list of genes to exclude from the features gene set}

\item{logist_loc}{the "location" parametr of the logistic function used to determine parametric distances between metacelles}

\item{logist_scale}{the "lscale" parametr of the logistic function used to determine parametric distances between metacelles}

\item{mgraph_id}{id of new object}

\item{mc_id}{meta cell id to work with}

\item{feats_gset}{gene set name for use for computing distances}
}
\value{
data.frame with mc1, mc2 and dist column
}
\description{
Compute metacell manifold graph using logistic distances and balanced K-nn
}
