% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgraph_logistic.R
\name{mgraph_comp_logist_mc2}
\alias{mgraph_comp_logist_mc2}
\title{Compute logistic distances between metacells}
\usage{
mgraph_comp_logist_mc2(
  mc_gene_mat,
  feature_genes,
  loc = 1,
  scale = 0.2,
  eps = 4e-05,
  max_d_fold = 3,
  max_deg = 4
)
}
\arguments{
\item{mc_gene_mat}{expression matrix metacells (rows) over genes (columns). mc2$X entry of anndata metacell2 object.}

\item{feature_genes}{feature genes used for logistic distance computation}

\item{loc}{location parameter of plogis function}

\item{scale}{scale parameter of plogis function}

\item{eps}{regularization used for logarithmic expression matrix log2(t(mc_gene_mat[,feature_genes] + eps)}

\item{max_d_fold}{threshold for cutoff of logistic distance compared to second-nearest neighbor}

\item{max_deg}{maximal degree of resulting manifold graph}
}
\value{
data.frame with mc1, mc2 and dist column
}
\description{
Compute logistic distances between metacells
}
