% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_add_mc2tnetflow}
\alias{scdb_add_mc2tnetflow}
\title{Add mc2tnetflow to single-cell database}
\usage{
scdb_add_mc2tnetflow(mcf, flow_id, scdb_dir)
}
\arguments{
\item{mcf}{- mc2tnetflow object}

\item{flow_id}{- id of mc2tnetflow}

\item{scdb_dir}{- directory of single-cell database}
}
\description{
Add mc2tnetflow to single-cell database
}
