% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_add_mc2tnetwork}
\alias{scdb_add_mc2tnetwork}
\title{Add mc2tnetwork to single-cell database}
\usage{
scdb_add_mc2tnetwork(mct, net_id, scdb_dir)
}
\arguments{
\item{mct}{- mc2tnetwork object}

\item{net_id}{- id of mc2tnetwork}

\item{scdb_dir}{- directory of single-cell database}
}
\description{
Add mc2tnetwork to single-cell database
}
