% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_add_mctnetflow}
\alias{scdb_add_mctnetflow}
\title{scdb_add_mctnetflow - add mctnetflow to the DB and cahce}
\usage{
scdb_add_mctnetflow(id, mctnetflow)
}
\arguments{
\item{id}{- id of mctnetflow}

\item{mctnetflow}{- mctnetflow data frame}
}
\description{
scdb_add_mctnetflow - add mctnetflow to the DB and cahce
}
