% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_add_mgraph2}
\alias{scdb_add_mgraph2}
\title{Add mgraph to single-cell database}
\usage{
scdb_add_mgraph2(mgraph, mgraph_id, scdb_dir)
}
\arguments{
\item{mgraph}{- mgraph object}

\item{mgraph_id}{- id of mgraph}

\item{scdb_dir}{- directory of single-cell database}
}
\description{
Add mgraph to single-cell database
}
