% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_del_mctnetflow}
\alias{scdb_del_mctnetflow}
\title{scdb_del_mctnetflow - del mctnetflow from the DB and cahce}
\usage{
scdb_del_mctnetflow(id)
}
\arguments{
\item{id}{- id of mctnetflow}
}
\description{
scdb_del_mctnetflow - del mctnetflow from the DB and cahce
}
