% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_flow_init}
\alias{scdb_flow_init}
\title{Initializing scdb_flow}
\usage{
scdb_flow_init()
}
\description{
This initialize scdb_flow to an already initialed scdb. This will extend the db with mctnetwork objects
}
\examples{
\dontrun{
# we first initialize a db object
scdb_flow_init()
net = scdb_mctnetwork("mynet")
}

}
