% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_mc2tnetflow}
\alias{scdb_mc2tnetflow}
\title{get mc2tnetflow object}
\usage{
scdb_mc2tnetflow(flow_id, scdb_dir)
}
\arguments{
\item{flow_id}{- id of mc2tnetflow}

\item{scdb_dir}{- directory of single-cell database}
}
\description{
get mc2tnetflow object
}
