% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_mc2tnetwork}
\alias{scdb_mc2tnetwork}
\title{get mc2tnetwork object}
\usage{
scdb_mc2tnetwork(net_id, scdb_dir)
}
\arguments{
\item{net_id}{- id of mc2tnetwork}

\item{scdb_dir}{- directory of single-cell database}
}
\description{
get mc2tnetwork object
}
