% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_mctnetflow}
\alias{scdb_mctnetflow}
\title{scdb_mctnetflow - get a mctnetflow object}
\usage{
scdb_mctnetflow(id)
}
\arguments{
\item{id}{- id of mctnetflow}
}
\description{
scdb_mctnetflow - get a mctnetflow object
}
