% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb_mcflows.r
\name{scdb_mgraph2}
\alias{scdb_mgraph2}
\title{Get mgraph from single-cell database}
\usage{
scdb_mgraph2(mgraph_id, scdb_dir)
}
\arguments{
\item{mgraph_id}{- id of mgraph}

\item{scdb_dir}{- directory of single-cell database}
}
\description{
Get mgraph from single-cell database
}
