% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\docType{class}
\name{tgMCTNetFlow-class}
\alias{tgMCTNetFlow-class}
\alias{tgMCTNetFlow}
\title{Flow solution over a metacell temporal network}
\description{
Relating to a mctnetwork object, this store on particular flow solution
}
\section{Slots}{

\describe{
\item{\code{net_id}}{- network model ID}

\item{\code{edge_flows}}{- flow of mincost solution}

\item{\code{mc_t_infer}}{- inferred fraction ofcells per mc and time}

\item{\code{mc_forward}}{- forward flow distribution (mc_forward[[t]][mc1,mc2] = how much of the flow from mc1 is going to mc2 at time [t,t+1]}

\item{\code{mc_backward}}{- backward flow distribution (mc_backward[[t]][mc1,mc2] = how much of the flow into mc1 came from  mc2 at time [t,t+1]}
}}

