% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetwork.r
\docType{class}
\name{tgMCTNetwork-class}
\alias{tgMCTNetwork-class}
\alias{tgMCTNetwork}
\title{temporal netwrok over metacells}
\description{
Splitting metacells over a discrete time axis, defining manifold connections and estimated flows over them
}
\section{Slots}{

\describe{
\item{\code{mc_t}}{distribution of metacells (rows) over time points (cols)}

\item{\code{mc_manif_cost}}{probability of moving between mcs}

\item{\code{mc_t_expansion_rate}}{matrix metacells over time group. Expansion/reduction rate per metacell and time bin. Can included effects of proliferation, ingress/egress & cell death}

\item{\code{temporal_bin_time}}{data.frame with columns bin and t giving for each temporal bin its actual time}

\item{\code{network}}{- a data frame defining the network structure}

\item{\code{metacell_names}}{- vector with metacell names}
}}

