% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctnetflow.r
\name{initialize,tgMCTNetFlow-method}
\alias{initialize,tgMCTNetFlow-method}
\title{Construct a flow object given an mctnetwork}
\usage{
\S4method{initialize}{tgMCTNetFlow}(.Object, net_id, flows = NULL)
}
\arguments{
\item{net_id}{metacell object id}

\item{flow}{named vector flows on each net edge (name is edge ID)}
}
\description{
Construct a flow object given an mctnetwork
}
